pyproject.toml
[project]
name = "assesment-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.3.0",
    "langchain-openrouter>=0.2.3",
    "langgraph>=1.2.0",
    "pydantic>=2.13.4",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.12",
]
